XXX: move these into a separate file and use the std edge
